Re-Build for Week 5
-Corrected Issue where Register & Login did not work

New Web Address: https://cis174-gtarr-web.azurewebsites.net/

+++ gTarrGames/gTarrGames.Domain/Entities/HighScore.cs
@@ -6,7 +6,7 @@ namespace gTarrGames.Domain.Entities
     {
         public Guid HighScoreId { get; set; }
         public Guid PersonId { get; set; }
-        public float Score { get; set; }
+        public decimal Score { get; set; }
         public DateTime DateAttained { get; set; }
     }
 }
