+++ gTarrGames/gTarrGames.Shared/ViewModels/HighScoreViewModel.cs
@@ -4,9 +4,8 @@ namespace gTarrGames.Shared.ViewModels
 {
     public class HighScoreViewModel
     {
-        
         public Guid PersonId { get; set; }
-        public float Score { get; set; }
+        public decimal Score { get; set; }
         public DateTime DateAttained { get; set; }
     }
 }
